refactor: changed category package name and source
