+++ src/main/java/com/amalitech/gpuconfigurator/controller/CategoryConfigController.java
@@ -3,7 +3,7 @@ package com.amalitech.gpuconfigurator.controller;
 import com.amalitech.gpuconfigurator.dto.GenericResponse;
 import com.amalitech.gpuconfigurator.dto.categoryconfig.CategoryConfigResponseDto;
 import com.amalitech.gpuconfigurator.dto.categoryconfig.CompatibleOptionDTO;
-import com.amalitech.gpuconfigurator.service.category.CategoryConfig.CategoryConfigService;
+import com.amalitech.gpuconfigurator.service.categoryConfig.CategoryConfigService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
